Formatted, restructured code and added TODOs
