+++ bookmyshow-notify/utils.py
@@ -1,0 +1,12 @@
+def validate_date(date):
+
+    # TODO: Validate the date format and compare
+    
+    return date
+
+
+def build_url(url, date):
+
+    # TODO: Get resultant URL from the passed arguments
+
+    return url
